+++ src/core/featuremodel.cpp
@@ -271,7 +271,7 @@ int FeatureModel::rowCount( const QModelIndex &parent ) const
   if ( parent.isValid() )
     return 0;
 
-  return mFeature.attributes().count();
+  return static_cast<int>( mFeature.attributes().count() );
 }
 
 QVariant FeatureModel::data( const QModelIndex &index, int role ) const
